Create BME280httpdTelegraph.ino
Some hacked together code which can be improved massively for sure.
But it's kind of my first attempt since a while to get this WEMOS D1 mini Pro to really be useful.
The driver here was that when the kids use all the hot water in their planet killing showers I don't know until I have only cold water.
The temps may need tweaking but my sensor is under lagging but never really exceeded 35C. I assume when there's no hot water left it's more like 20 or 15 instead. May change the sensor and its location in the future and need to tweak the values.
Will likely tidy up this abortion of a program too.